Revamped registration process so registered users don't need to reregister
The new process is:

1. Enter email
2. If email is not registered, go to registration page, else if email
   is registered, go to downloads page.
