+++ APSIM.Registration.Portal/APSIM.Registration.Portal.csproj
@@ -69,9 +69,12 @@
   <ItemGroup>
     <Content Include="ApsimWebStyle.css" />
     <Content Include="APSIM_NonCommercial_RD_licence.htm" />
+    <Content Include="Downloads.aspx" />
     <Content Include="EmailBody.html" />
-    <Content Include="OtherDisclaimer.html" />
     <Content Include="Main.aspx" />
+    <Content Include="OtherDisclaimer.html" />
+    <Content Include="Register.aspx" />
+    <Content Include="Register.js" />
     <Content Include="Screen_Styles.css" />
     <Content Include="APSIM_NonCommercial_RD_licence.pdf" />
     <Content Include="Thankyou.aspx" />
@@ -87,12 +90,28 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Downloads.aspx.cs">
+      <DependentUpon>Downloads.aspx</DependentUpon>
+      <SubType>ASPXCodeBehind</SubType>
+    </Compile>
+    <Compile Include="Downloads.aspx.designer.cs">
+      <DependentUpon>Downloads.aspx</DependentUpon>
+    </Compile>
     <Compile Include="Main.aspx.cs">
       <DependentUpon>Main.aspx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
     <Compile Include="Main.aspx.designer.cs">
       <DependentUpon>Main.aspx</DependentUpon>
+    </Compile>
+    <Compile Include="Product.cs" />
+    <Compile Include="ProductVersion.cs" />
+    <Compile Include="Register.aspx.cs">
+      <DependentUpon>Register.aspx</DependentUpon>
+      <SubType>ASPXCodeBehind</SubType>
+    </Compile>
+    <Compile Include="Register.aspx.designer.cs">
+      <DependentUpon>Register.aspx</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Thankyou.aspx.cs">
